subterminator/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "subterminator"
version = "0.1.0"
description = "CLI tool for automating subscription cancellations"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}
authors = [
    {name = "SubTerminator Team"}
]
dependencies = [
    "playwright>=1.58",
    "typer[all]>=0.21",
    "questionary>=2.0.0",
    "anthropic>=0.40",
    "pydantic>=2.0",
    "requests",
    "pillow>=12.1.0",
    # MCP orchestrator dependencies
    "mcp>=1.0.0",
    "langchain-core>=0.3.0",
    "langchain-anthropic>=0.3.0",
    "langchain-openai>=0.3.0",
]

[dependency-groups]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
    "pytest-cov>=5.0",
    "pytest-mock>=3.0",
    "mypy>=1.8",
    "ruff>=0.5",
    "python-dotenv>=1.0.0",
    "pyyaml>=6.0.3",
]

[project.scripts]
subterminator = "subterminator.cli.main:app"

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
line-length = 88
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[tool.ruff.lint.per-file-ignores]
# LLM prompts and system prompts have inherently long lines
"src/subterminator/mcp_orchestrator/task_runner.py" = ["E501"]
"src/subterminator/mcp_orchestrator/services/netflix.py" = ["E501"]

[tool.mypy]
python_version = "3.11"
strict = false
disallow_untyped_defs = true
check_untyped_defs = true


[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
addopts = "-v --cov=subterminator --cov-report=term-missing"
